editor49

+++ text_editor/editor.c
@@ -53,7 +53,7 @@ void handle_write_command(Document *document, const char *command) {
 			// printf("end: %c\n", *end);
 			res = malloc(length + 1);
 			strncpy(res, start, length);
-			if(line_num > Document_size(document)){
+			if((size_t)line_num > Document_size(document)){
 				Document_insert_line(document, line_num++, res);
 			}
 			else Document_set_line(document, line_num++, res);
